# foundation/docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_upload_capture_file_route_api_v1_captures_files_post": {
        "properties": {
          "domain_hint": {
            "anyOf": [
              {
                "maxLength": 80,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Domain Hint"
          },
          "file": {
            "format": "binary",
            "title": "File",
            "type": "string"
          },
          "sensitivity": {
            "default": "internal",
            "enum": [
              "public",
              "internal",
              "confidential",
              "restricted"
            ],
            "title": "Sensitivity",
            "type": "string"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 300,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_capture_file_route_api_v1_captures_files_post",
        "type": "object"
      },
      "CaptureCreate": {
        "properties": {
          "content": {
            "maxLength": 500000,
            "minLength": 1,
            "title": "Content",
            "type": "string"
          },
          "domain_hint": {
            "anyOf": [
              {
                "maxLength": 80,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Domain Hint"
          },
          "sensitivity": {
            "default": "internal",
            "enum": [
              "public",
              "internal",
              "confidential",
              "restricted"
            ],
            "title": "Sensitivity",
            "type": "string"
          },
          "title": {
            "maxLength": 300,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "type": {
            "default": "conversation",
            "enum": [
              "conversation",
              "note",
              "markdown",
              "url",
              "repository"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "content"
        ],
        "title": "CaptureCreate",
        "type": "object"
      },
      "CaptureListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CaptureResponse"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "CaptureListResponse",
        "type": "object"
      },
...
      "CaptureResponse": {
        "properties": {
          "checksum_sha256": {
            "title": "Checksum Sha256",
            "type": "string"
          },
          "classification": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Classification"
          },
          "content": {
            "title": "Content",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "domain_hint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Domain Hint"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "review_status": {
            "title": "Review Status",
            "type": "string"
          },
          "sensitivity": {
            "title": "Sensitivity",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "workspace_id": {
            "format": "uuid",
            "title": "Workspace Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "workspace_id",
          "type",
          "title",
          "content",
          "domain_hint",
          "sensitivity",
          "status",
          "review_status",
          "summary",
          "classification",
          "checksum_sha256",
          "created_at",
          "updated_at"
        ],
        "title": "CaptureResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Governed execution API for Ryan Agent OS.",
    "title": "Ryan Agent OS API",
    "version": "0.4.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/captures": {
      "get": {
        "operationId": "list_capture_route_api_v1_captures_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaptureListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Capture Route",
        "tags": [
          "captures"
        ]
      },
      "post": {
        "operationId": "create_capture_route_api_v1_captures_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaptureCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaptureResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Capture Route",
        "tags": [
          "captures"
        ]
      }
    },
    "/api/v1/captures/files": {
      "post": {
        "operationId": "upload_capture_file_route_api_v1_captures_files_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_capture_file_route_api_v1_captures_files_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileCaptureResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upload Capture File Route",
        "tags": [
          "captures"
        ]
      }
    },
    "/api/v1/captures/{capture_id}": {
      "get": {
        "operationId": "get_capture_route_api_v1_captures__capture_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "capture_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Capture Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaptureResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Capture Route",
        "tags": [
          "captures"
        ]
      }
    },
    "/api/v1/captures/{capture_id}/archive": {
      "post": {
        "operationId": "archive_capture_route_api_v1_captures__capture_id__archive_post",
        "parameters": [
          {
            "in": "path",
            "name": "capture_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Capture Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaptureResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Archive Capture Route",
        "tags": [
          "captures"
        ]
      }
    },
    "/api/v1/captures/{capture_id}/process": {
      "post": {
        "operationId": "process_capture_route_api_v1_captures__capture_id__process_post",
        "parameters": [
          {
            "in": "path",
            "name": "capture_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Capture Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaptureResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Process Capture Route",
        "tags": [
          "captures"
        ]
      }
    },
    "/api/v1/captures/{capture_id}/projects": {
      "post": {
        "operationId": "create_project_from_capture_route_api_v1_captures__capture_id__projects_post",
        "parameters": [
          {
            "in": "path",
            "name": "capture_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Capture Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectFromCaptureRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Project From Capture Route",
        "tags": [
          "captures"
        ]
      }
    },
    "/api/v1/captures/{capture_id}/relations": {
      "post": {
        "operationId": "create_capture_relation_route_api_v1_captures__capture_id__relations_post",
        "parameters": [
          {
            "in": "path",
            "name": "capture_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Capture Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaptureRelationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaptureRelationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Capture Relation Route",
        "tags": [
          "captures"
        ]
      }
    },
    "/api/v1/captures/{capture_id}/retry": {
      "post": {
        "operationId": "retry_capture_route_api_v1_captures__capture_id__retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "capture_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Capture Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileCaptureResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Retry Capture Route",
        "tags": [
          "captures"
        ]
      }
    },
    "/api/v1/captures/{capture_id}/review": {
      "get": {
        "operationId": "get_capture_review_route_api_v1_captures__capture_id__review_get",
        "parameters": [
          {
            "in": "path",
            "name": "capture_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Capture Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaptureReviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Capture Review Route",
        "tags": [
          "captures"
        ]
      }
    },
...
    "/api/v1/projects/{project_id}/captures/{capture_id}": {
      "post": {
        "operationId": "link_capture_route_api_v1_projects__project_id__captures__capture_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "capture_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Capture Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Link Capture Route",
        "tags": [
          "projects"
        ]
      }
    },
...
  }
}